+++ app/imports/ui/pages/UserProfileById.jsx
@@ -1,13 +1,20 @@
 import React from 'react';
-import { Grid, Loader, Header, Container, Image, Icon, Card, Rating } from 'semantic-ui-react';
+import { Grid, Loader, Header, Container, Image, Icon, Card, Rating, Item } from 'semantic-ui-react';
 import { Users } from '/imports/api/user/user';
 import { Meteor } from 'meteor/meteor';
 import { withTracker } from 'meteor/react-meteor-data';
 import PropTypes from 'prop-types';
-import UserShowCase from '../components/UserProfileShowcase';
+import { sortBy } from 'underscore';
+import ShowcaseItem from '../components/ShowcaseItem';
+import { Items } from '../../api/item/item';
 
 /** Renders the Page for editing a single document. */
 class UserProfileById extends React.Component {
+
+  getItems(items, owner) {
+    const stuff = sortBy(items, 'owner');
+    return stuff.filter(item => item.owner === owner).map((item) => <ShowcaseItem key={item._id} item={item}/>);
+  }
 
   /** If the subscription(s) have been received, render the page, otherwise show a loading icon. */
   render() {
@@ -21,6 +28,7 @@ class UserProfileById extends React.Component {
     const borderStyle = { border: 'solid 1px #feffff' };
     const cardColor = { backgroundColor: '#feffff' };
     const showcaseRow = { marginTop: '64px' };
+    const showcaseStyle = { marginTop: '8px', marginBottom: '128px' };
     return (
         <Grid container verticalAlign='middle' style={gridStyle}>
           <style>{'body { background: url(images/uh-logo.png) no-repeat center fixed; }'}</style>
@@ -48,7 +56,7 @@ class UserProfileById extends React.Component {
                       </a>
                     </Card.Content>
                     <Card.Content>
-                      <Rating icon='star' defaultRating={4} maxRating={5} />
+                      <Rating icon='star' defaultRating={4} maxRating={5}/>
                     </Card.Content>
                   </Card>
                 </Grid.Column>
@@ -57,7 +65,20 @@ class UserProfileById extends React.Component {
 
             <Grid>
               <Grid.Row style={showcaseRow}>
-                <UserShowCase/>
+                <Grid container centered style={showcaseStyle}>
+
+                  <Card fluid style={cardColor}>
+                    <Card.Content>
+                      <Card.Header style={cardFontStyle}><Icon name='warehouse' circular/>The Goods</Card.Header>
+                    </Card.Content>
+                    <Card.Content>
+                      <Item.Group>
+                        {this.getItems(this.props.item, this.props.doc.username)}
+                      </Item.Group>
+                    </Card.Content>
+                  </Card>
+
+                </Grid>
               </Grid.Row>
             </Grid>
           </Container>
@@ -71,6 +92,7 @@ class UserProfileById extends React.Component {
 UserProfileById.propTypes = {
   doc: PropTypes.array,
   ready: PropTypes.bool.isRequired,
+  item: PropTypes.object.isRequired,
 };
 
 /** withTracker connects Meteor data to React components. https://guide.meteor.com/react.html#using-withTracker */
@@ -80,6 +102,7 @@ export default withTracker(({ match }) => {
   // Get access to Stuff documents.
   const subscription = Meteor.subscribe('UserSearch');
   return {
+    item: Items.find({}).fetch(),
     doc: Users.findOne(documentId),
     ready: subscription.ready(),
   };
